Move MarvelFeedProvider & MainComposer to Production
